Refrain from changing presence in `on_ready`
Discord has a high chance to completely disconnect during the READY or GUILD_CREATE events (1006 close code) and there is nothing to do to prevent it. 
As noted in the docs, `on_ready` is also triggered multiple times, not just once.
Therefore it is highly inadvisable to change presence in `on_ready`
This pr fixes that and puts `activity=` in client constructor instead.